pyproject.toml
[project]
name = "customer-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "mypy>=1.17.0",
    "openai-agents>=0.2.4",
]

[project.scripts]
customer-support = "customer_support:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
